refactor: move log page and navbar tests to test_scripts_helpers.py
Brings test_scripts_api.py under 500 lines (468).
