+++ tests/unit/test_scripts_helpers.py
@@ -1,19 +1,55 @@
 """tests/unit/test_scripts_helpers.py
 
-Tests for script discovery, validation, and history helper functions.
+Tests for script discovery, validation, history helpers, page routes, and navbar.
 """
 
 from __future__ import annotations
 
+import json
 from pathlib import Path
+from typing import Any
+from unittest.mock import patch
 
 import pytest
+
+from dango.auth.models import Role, User
 
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
 
 _PATCH_HELPERS = "dango.web.routes.scripts_helpers"
+_PATCH_SCRIPTS = "dango.web.routes.scripts"
+
+
+def _make_admin_user() -> User:
+    return User(id="admin-id", email="admin@test.com", role=Role.ADMIN, is_active=True)
+
+
+def _make_app(tmp_path: Path, user: User | None = None) -> Any:
+    """Build a minimal FastAPI app with the scripts router and injected user."""
+    from fastapi import FastAPI
+    from fastapi.responses import JSONResponse
+
+    from dango.exceptions import AuthorizationError
+    from dango.web.routes.scripts import router
+
+    app = FastAPI()
+    app.state.project_root = tmp_path
+
+    test_user = user or _make_admin_user()
+
+    @app.middleware("http")
+    async def inject_user(request: Any, call_next: Any) -> Any:
+        request.state.user = test_user
+        return await call_next(request)
+
+    @app.exception_handler(AuthorizationError)
+    async def auth_error_handler(request: Any, exc: AuthorizationError) -> Any:
+        return JSONResponse(status_code=403, content={"detail": str(exc)})
+
+    app.include_router(router)
+    return app
 
 
 def _write_script(scripts_dir: Path, name: str, content: str = "print('hello')") -> Path:
@@ -222,3 +258,116 @@ class TestHistoryHelpers:
 
         result = _load_history(tmp_path, "nonexistent.py")
         assert result == []
+
+
+# ---------------------------------------------------------------------------
+# TestScriptLogPage
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.unit
+class TestScriptLogPage:
+    """Tests for GET /scripts/{name}/logs/{run_id}."""
+
+    def test_renders_with_content(self, tmp_path: Path):
+        """Renders log page with stdout/stderr content."""
+        from starlette.testclient import TestClient
+
+        import dango.web.routes.scripts_helpers as helpers
+
+        _write_script(tmp_path / "scripts", "hello.py")
+
+        log_dir = helpers._get_log_dir(tmp_path) / "test-run-id"
+        log_dir.mkdir(parents=True, exist_ok=True)
+        (log_dir / "stdout.txt").write_text("hello world\n")
+        (log_dir / "stderr.txt").write_text("warning: something\n")
+        (log_dir / "meta.json").write_text(
+            json.dumps(
+                {
+                    "run_id": "test-run-id",
+                    "script_name": "hello.py",
+                    "started_at": "2026-01-01T00:00:00+00:00",
+                    "finished_at": "2026-01-01T00:00:05+00:00",
+                    "duration_seconds": 5.0,
+                    "status": "success",
+                    "exit_code": 0,
+                    "error": None,
+                }
+            )
+        )
+
+        app = _make_app(tmp_path)
+        client = TestClient(app)
+
+        with patch(f"{_PATCH_SCRIPTS}.get_project_root", return_value=tmp_path):
+            resp = client.get("/scripts/hello.py/logs/test-run-id")
+        assert resp.status_code == 200
+        assert "hello world" in resp.text
+        assert "warning: something" in resp.text
+        assert "test-run-id" in resp.text
+
+    def test_non_existent_run_returns_404(self, tmp_path: Path):
+        """Returns 404 page when run doesn't exist."""
+        from starlette.testclient import TestClient
+
+        _write_script(tmp_path / "scripts", "hello.py")
+
+        app = _make_app(tmp_path)
+        client = TestClient(app)
+
+        with patch(f"{_PATCH_SCRIPTS}.get_project_root", return_value=tmp_path):
+            resp = client.get("/scripts/hello.py/logs/nonexistent-run")
+        assert resp.status_code == 404
+        assert "not found" in resp.text.lower()
+
+    def test_non_existent_script_returns_404(self, tmp_path: Path):
+        """Returns 404 page when script doesn't exist."""
+        from starlette.testclient import TestClient
+
+        app = _make_app(tmp_path)
+        client = TestClient(app)
+
+        with patch(f"{_PATCH_SCRIPTS}.get_project_root", return_value=tmp_path):
+            resp = client.get("/scripts/nonexistent.py/logs/some-run")
+        assert resp.status_code == 404
+        assert "not found" in resp.text.lower()
+
+    def test_handles_empty_stdout_stderr(self, tmp_path: Path):
+        """Renders gracefully when stdout/stderr are empty."""
+        from starlette.testclient import TestClient
+
+        import dango.web.routes.scripts_helpers as helpers
+
+        _write_script(tmp_path / "scripts", "hello.py")
+
+        log_dir = helpers._get_log_dir(tmp_path) / "empty-run"
+        log_dir.mkdir(parents=True, exist_ok=True)
+
+        app = _make_app(tmp_path)
+        client = TestClient(app)
+
+        with patch(f"{_PATCH_SCRIPTS}.get_project_root", return_value=tmp_path):
+            resp = client.get("/scripts/hello.py/logs/empty-run")
+        assert resp.status_code == 200
+        assert "(empty)" in resp.text
+
+
+# ---------------------------------------------------------------------------
+# TestNavBar
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.unit
+class TestNavBar:
+    """Tests that base.html contains Scripts nav link."""
+
+    def test_base_html_contains_scripts_link(self):
+        """Base template contains Scripts nav link."""
+        from pathlib import Path
+
+        base_path = (
+            Path(__file__).parent.parent.parent / "dango" / "web" / "templates" / "base.html"
+        )
+        content = base_path.read_text()
+        assert 'href="/scripts"' in content
+        assert "Scripts" in content
